lambda/arquitecto/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
python-docx>=0.8.11
pandas>=1.5.0
openpyxl>=3.0.10
PyYAML>=6.0
requests>=2.28.0
